+++ packages/palette/src/elements/CleanTag/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CleanTag"
