--- packages/palette/src/elements/Tag/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Tag"
