+++ packages/palette/src/elements/index.tsx
@@ -39,7 +39,7 @@ export * from "./StackableBorderBox"
 export * from "./StaticCountdownTimer"
 export * from "./Stepper"
 export * from "./Tabs"
-export * from "./Tag"
+export * from "./CleanTag"
 export * from "./TextArea"
 export * from "./TimeRemaining"
 export * from "./Toggle"
